[LEMS-2737/move-scoring-logic] WIP: getSupportedAlignments
