+++ packages/perseus-editor/src/diffs/renderer-diff.tsx
@@ -2,13 +2,12 @@ import {Widgets} from "@khanacademy/perseus";
 /**
  * A side by side diff view for Perseus renderers.
  */
+import {WidgetLogic, type PerseusRenderer} from "@khanacademy/perseus-core";
 import * as React from "react";
 import _ from "underscore";
 
 import TextDiff from "./text-diff";
 import WidgetDiff from "./widget-diff";
-
-import type {PerseusRenderer} from "@khanacademy/perseus-core";
 
 // In diffs, only show the widgetInfo props that can change
 const filterWidgetInfo = function (widgetInfo, showAlignmentOptions: boolean) {
@@ -19,7 +18,7 @@ const filterWidgetInfo = function (widgetInfo, showAlignmentOptions: boolean) {
     // Show alignment options iff multiple valid ones exist for this widget
     if (
         showAlignmentOptions &&
-        Widgets.getSupportedAlignments(type).length > 1
+        WidgetLogic.getSupportedAlignments(type).length > 1
     ) {
         // @ts-expect-error - TS2339 - Property 'alignment' does not exist on type '{ readonly options: any; }'.
         filteredWidgetInfo.alignment = alignment;
